+++ Changelog.md
@@ -40,7 +40,7 @@
 * Fix not closing file in `LogManager#read_log`
 * Fix booking already booked server by someone
 * Fix leftovers of Histories after deleting them
-* Fix critical stop on `read_progress` `Errno::ECONNRESET`, `Errno::ECONNREFUSED` exception
+* Fix critical stop on `read_progress` `Errno::ECONNRESET`, `Errno::ECONNREFUSED`, `Errno::EHOSTUNREACH` exception
 * Fix critical crash if log file for server is lost
 
 ## 1.15.0